Reorganise the organise module into submodules
With tests and various helper functions, the organise module was too
big. By adding new candidate and confirmed submodules the size of each
individual module is reduced and the code is logically organised based
on which chain it operates on.
