+++ p2poolv2_lib/src/store/organise/mod.rs
@@ -14,24 +14,12 @@
 // You should have received a copy of the GNU General Public License along with
 // P2Poolv2. If not, see <https://www.gnu.org/licenses/>.
 
-use super::{
-    ColumnFamily, Store,
-    block_tx_metadata::{BlockMetadata, Status},
-    writer::StoreError,
-};
-use bitcoin::{
-    BlockHash, Work,
-    consensus::{self, encode},
-};
-use std::collections::VecDeque;
-use tracing::debug;
-pub mod organise_share;
+use super::{ColumnFamily, Store, writer::StoreError};
+use bitcoin::{BlockHash, Work, consensus::encode};
 
-const CANDIDATE_SUFFIX: &str = ":c";
-const CONFIRMED_SUFFIX: &str = ":f";
-const TOP_CANDIDATE_KEY: &str = "meta:top_candidate_height";
-const TOP_CONFIRMED_KEY: &str = "meta:top_confirmed_height";
-const BRANCH_INITIAL_CAPACITY: usize = 16;
+mod candidate;
+mod confirmed;
+pub mod organise_share;
 
 /// Type to capture candidate and confirmed chains as vector of
 /// height, blockhash pairs
@@ -49,255 +37,21 @@ pub(crate) struct TopResult {
 }
 
 /// Returns key for height with provided suffix
-fn height_to_key_with_suffix(height: Height, suffix: &str) -> Vec<u8> {
+pub(super) fn height_to_key_with_suffix(height: Height, suffix: &str) -> Vec<u8> {
     [&height.to_be_bytes(), suffix.as_bytes()].concat()
 }
 
 impl Store {
-    /// Increment top candidate key if height is one more than current height
-    ///
-    /// Only updates top if it is more than one higher.
-    fn increment_top_candidate(
-        &self,
-        height: Height,
-        batch: &mut rocksdb::WriteBatch,
-    ) -> Result<Height, StoreError> {
-        let block_height_cf = self.db.cf_handle(&ColumnFamily::BlockHeight).unwrap();
-
-        let use_height = match self.get_top_candidate_height() {
-            Ok(current_top_height) => {
-                if height.saturating_sub(current_top_height) == 1 {
-                    height
-                } else {
-                    return Err(StoreError::Database("Mismatch in top height".into()));
-                }
-            }
-            Err(StoreError::NotFound(_reason)) => height, // Use share height if no candidate top present
-            Err(e) => return Err(e),
-        };
-        let serialized_height = consensus::serialize(&use_height);
-        batch.put_cf(
-            &block_height_cf,
-            TOP_CANDIDATE_KEY.as_bytes().as_ref(),
-            serialized_height,
-        );
-        Ok(use_height)
-    }
-
-    /// Directly set top candidate height without consecutive-height validation.
-    ///
-    /// Used by `reorg_candidate` which computes the correct final height
-    /// locally instead of reading stale DB state within a single WriteBatch.
-    fn set_top_candidate_height(&self, height: Height, batch: &mut rocksdb::WriteBatch) {
-        let block_height_cf = self.db.cf_handle(&ColumnFamily::BlockHeight).unwrap();
-        let serialized_height = consensus::serialize(&height);
-        batch.put_cf(
-            &block_height_cf,
-            TOP_CANDIDATE_KEY.as_bytes().as_ref(),
-            serialized_height,
-        );
-    }
-
-    /// Delete top candidate height
-    /// Used when entire candidate chain has been moved to confirmed chain
-    fn delete_top_candidate_height(&self, batch: &mut rocksdb::WriteBatch) {
-        let block_height_cf = self.db.cf_handle(&ColumnFamily::BlockHeight).unwrap();
-        batch.delete_cf(&block_height_cf, TOP_CANDIDATE_KEY.as_bytes().as_ref());
-    }
-
-    /// Get top candidate height from candidates index
-    pub(crate) fn get_top_candidate_height(&self) -> Result<Height, StoreError> {
-        let block_height_cf = self.db.cf_handle(&ColumnFamily::BlockHeight).unwrap();
-        match self
-            .db
-            .get_cf(&block_height_cf, TOP_CANDIDATE_KEY.as_bytes().as_ref())
-        {
-            Ok(Some(height_bytes)) => Ok(encode::deserialize(&height_bytes)?),
-            Ok(None) => Err(StoreError::NotFound("No candidate found at top".into())),
-            Err(e) => Err(e.into()),
-        }
-    }
-
-    /// Write a candidate index entry directly into the batch.
-    fn put_candidate_entry(
-        &self,
-        height: Height,
-        blockhash: &BlockHash,
-        batch: &mut rocksdb::WriteBatch,
-    ) {
-        let block_height_cf = self.db.cf_handle(&ColumnFamily::BlockHeight).unwrap();
-        let key = height_to_key_with_suffix(height, CANDIDATE_SUFFIX);
-        let serialized = consensus::serialize(blockhash);
-        batch.put_cf(&block_height_cf, key, serialized);
-    }
-
-    /// Delete a candidate index entry from the batch.
-    fn delete_candidate_entry(&self, height: Height, batch: &mut rocksdb::WriteBatch) {
-        let block_height_cf = self.db.cf_handle(&ColumnFamily::BlockHeight).unwrap();
-        let key = height_to_key_with_suffix(height, CANDIDATE_SUFFIX);
-        batch.delete_cf(&block_height_cf, key);
-    }
-
-    /// Write a confirmed index entry directly into the batch.
-    fn put_confirmed_entry(
-        &self,
-        height: Height,
-        blockhash: &BlockHash,
-        batch: &mut rocksdb::WriteBatch,
-    ) {
-        let block_height_cf = self.db.cf_handle(&ColumnFamily::BlockHeight).unwrap();
-        let key = height_to_key_with_suffix(height, CONFIRMED_SUFFIX);
-        let serialized = consensus::serialize(blockhash);
-        batch.put_cf(&block_height_cf, key, serialized);
-    }
-
-    /// Delete a confirmed index entry from the batch.
-    fn delete_confirmed_entry(&self, height: Height, batch: &mut rocksdb::WriteBatch) {
-        let block_height_cf = self.db.cf_handle(&ColumnFamily::BlockHeight).unwrap();
-        let key = height_to_key_with_suffix(height, CONFIRMED_SUFFIX);
-        batch.delete_cf(&block_height_cf, key);
-    }
-
-    /// Set top confirmed height
-    /// The required height checks are already made in make_confirmed
-    pub(crate) fn set_top_confirmed_height(&self, height: Height, batch: &mut rocksdb::WriteBatch) {
-        let block_height_cf = self.db.cf_handle(&ColumnFamily::BlockHeight).unwrap();
-        let serialized_height = consensus::serialize(&height);
-        batch.put_cf(
-            &block_height_cf,
-            TOP_CONFIRMED_KEY.as_bytes().as_ref(),
-            serialized_height,
-        );
-    }
-
-    /// Get top confirmed height from confirmed index
-    pub(crate) fn get_top_confirmed_height(&self) -> Result<Height, StoreError> {
-        let block_height_cf = self.db.cf_handle(&ColumnFamily::BlockHeight).unwrap();
-        match self
-            .db
-            .get_cf(&block_height_cf, TOP_CONFIRMED_KEY.as_bytes().as_ref())
-        {
-            Ok(Some(height_bytes)) => Ok(encode::deserialize(&height_bytes)?),
-            Ok(None) => Err(StoreError::NotFound("No confirmed found at top".into())),
-            Err(e) => Err(e.into()),
-        }
-    }
-
-    /// Get top candidate after looking up top candidate height
-    pub(crate) fn get_top_candidate(&self) -> Result<TopResult, StoreError> {
-        let height = self.get_top_candidate_height()?;
-        let hash = self.get_candidate_at_height(height)?;
-        let metadata = self.get_block_metadata(&hash)?;
-        Ok(TopResult {
-            hash,
-            height,
-            work: metadata.chain_work,
-        })
-    }
-
-    /// Get top confirmed after looking up top confirmed height
-    pub(crate) fn get_top_confirmed(&self) -> Result<TopResult, StoreError> {
-        let height = self.get_top_confirmed_height()?;
-        let hash = self.get_confirmed_at_height(height)?;
-        let metadata = self.get_block_metadata(&hash)?;
-        Ok(TopResult {
-            hash,
-            height,
-            work: metadata.chain_work,
-        })
-    }
-
-    /// Add blockhash as a candidate at provided height.
-    ///
-    /// Also Updates the metadata status to Candidate.
-    pub(crate) fn append_to_candidates(
-        &self,
-        blockhash: &BlockHash,
-        height: Height,
-        metadata: &mut BlockMetadata,
-        batch: &mut rocksdb::WriteBatch,
-    ) -> Result<Option<Height>, StoreError> {
-        let block_height_cf = self.db.cf_handle(&ColumnFamily::BlockHeight).unwrap();
-        let key = height_to_key_with_suffix(height, CANDIDATE_SUFFIX);
-
-        let serialized_blockhash = consensus::serialize(blockhash);
-        batch.put_cf(&block_height_cf, key, serialized_blockhash);
-
-        self.increment_top_candidate(height, batch)?;
-
-        metadata.status = Status::Candidate;
-        self.update_block_metadata(blockhash, metadata, batch)?;
-        Ok(Some(height))
-    }
-
-    /// Get list of (height, blockhash) pairs from given blockhash up to top candidate.
-    /// The blockhash is known to be on the candidates chain.
-    pub(crate) fn get_candidates_chain(
-        &self,
-        blockhash: &BlockHash,
-        top_candidate: Option<&TopResult>,
-    ) -> Result<Chain, StoreError> {
-        let Ok(metadata) = self.get_block_metadata(blockhash) else {
-            return Err(StoreError::NotFound(
-                "Block metadata not found for branch point".into(),
-            ));
-        };
-        let Some(height) = metadata.expected_height else {
-            return Err(StoreError::NotFound(
-                "Block metadata doesn't have an expected height".into(),
-            ));
-        };
-        let Some(top) = top_candidate else {
-            return Err(StoreError::NotFound(
-                "No top candidate height found when reorging candidate chain".into(),
-            ));
-        };
-        self.get_candidates(height, top.height)
-    }
-
-    /// Fetch a list of (height, blockhash) pairs on the candidates chain between
-    /// the given heights, inclusive.
-    pub(crate) fn get_candidates(&self, from: Height, to: Height) -> Result<Chain, StoreError> {
-        debug!("Get candidates from {from} to {to}");
-        self.get_chain_range(from, to, CANDIDATE_SUFFIX)
-    }
-
-    /// Get list of (height, blockhash) pairs from given blockhash up to top confirmed.
-    /// The blockhash is known to be on the confirmed chain.
-    pub(crate) fn get_confirmed_chain(
-        &self,
-        blockhash: &BlockHash,
-        top_confirmed: Option<&TopResult>,
-    ) -> Result<Chain, StoreError> {
-        let Ok(metadata) = self.get_block_metadata(blockhash) else {
-            return Err(StoreError::NotFound(
-                "Block metadata not found for branch point".into(),
-            ));
-        };
-        let Some(height) = metadata.expected_height else {
-            return Err(StoreError::NotFound(
-                "Block metadata doesn't have an expected height".into(),
-            ));
-        };
-        let Some(top) = top_confirmed else {
-            return Err(StoreError::NotFound(
-                "No top confirmed height found when fetching confirmed chain".into(),
-            ));
-        };
-        self.get_confirmed(height, top.height)
-    }
-
-    /// Fetch a list of (height, blockhash) pairs on the confirmed chain between
-    /// the given heights, inclusive.
-    pub(crate) fn get_confirmed(&self, from: Height, to: Height) -> Result<Chain, StoreError> {
-        self.get_chain_range(from, to, CONFIRMED_SUFFIX)
-    }
-
     /// Fetch blockhashes from the BlockHeight CF for a given height range
     /// and key suffix. Filters iterator results to only include keys whose
     /// suffix matches, avoiding cross-contamination between candidate (":c")
     /// and confirmed (":f") entries that share the same height prefix.
-    fn get_chain_range(&self, from: Height, to: Height, suffix: &str) -> Result<Chain, StoreError> {
+    pub(super) fn get_chain_range(
+        &self,
+        from: Height,
+        to: Height,
+        suffix: &str,
+    ) -> Result<Chain, StoreError> {
         if from > to {
             return Ok(Vec::new());
         }
@@ -328,1082 +82,5 @@ impl Store {
             }
         }
         Ok((from..=to).zip(results).collect())
-    }
-
-    /// Add blockhash as a confirmed at provided height.
-    ///
-    /// Only adds to the confirmed index if the height is one more than the
-    /// current top confirmed height (or if there is no top yet).
-    ///
-    /// Also updates the metadata status to Confirmed.
-    ///
-    /// Returns error if height check fails.
-    pub(crate) fn append_to_confirmed(
-        &self,
-        blockhash: &BlockHash,
-        height: Height,
-        metadata: &mut BlockMetadata,
-        batch: &mut rocksdb::WriteBatch,
-    ) -> Result<(), StoreError> {
-        let current_top = match self.get_top_confirmed_height() {
-            Ok(top) => top,
-            Err(StoreError::NotFound(_)) => 0,
-            Err(e) => return Err(e),
-        };
-
-        // Only add if this is the first entry or height is exactly one more than current top
-        // Or if it is the first confirmation
-        if height.saturating_sub(current_top) != 1 && !(height == 0 && current_top == 0) {
-            return Err(StoreError::Database("Incorrect confirmation".into()));
-        }
-
-        let block_height_cf = self.db.cf_handle(&ColumnFamily::BlockHeight).unwrap();
-        let key = height_to_key_with_suffix(height, CONFIRMED_SUFFIX);
-
-        let serialized_blockhash = consensus::serialize(blockhash);
-        batch.put_cf(&block_height_cf, key, serialized_blockhash);
-
-        self.set_top_confirmed_height(height, batch);
-
-        metadata.status = Status::Confirmed;
-        self.update_block_metadata(blockhash, metadata, batch)?;
-        Ok(())
-    }
-
-    /// Get the candidate blockhash at a specific height
-    pub fn get_candidate_at_height(&self, height: Height) -> Result<BlockHash, StoreError> {
-        let block_height_cf = self.db.cf_handle(&ColumnFamily::BlockHeight).unwrap();
-        let key = height_to_key_with_suffix(height, CANDIDATE_SUFFIX);
-
-        match self.db.get_cf::<&[u8]>(&block_height_cf, key.as_ref()) {
-            Ok(Some(blockhash_bytes)) => Ok(encode::deserialize(&blockhash_bytes)?),
-            Ok(None) => Err(StoreError::NotFound(format!(
-                "No candidate found at height {height}"
-            ))),
-            Err(e) => Err(e.into()),
-        }
-    }
-
-    /// Get the confirmed blockhash at a specific height
-    pub fn get_confirmed_at_height(&self, height: Height) -> Result<BlockHash, StoreError> {
-        let block_height_cf = self.db.cf_handle(&ColumnFamily::BlockHeight).unwrap();
-        let key = height_to_key_with_suffix(height, CONFIRMED_SUFFIX);
-
-        match self.db.get_cf::<&[u8]>(&block_height_cf, key.as_ref()) {
-            Ok(Some(blockhash_bytes)) => Ok(encode::deserialize(&blockhash_bytes)?),
-            Ok(None) => Err(StoreError::NotFound(format!(
-                "No confirmed found at height {height}"
-            ))),
-            Err(e) => Err(e.into()),
-        }
-    }
-
-    /// Check if a blockhash has Confirmed status in its metadata.
-    pub fn is_confirmed(&self, blockhash: &BlockHash) -> bool {
-        self.get_block_metadata(blockhash)
-            .map(|m| m.status == Status::Confirmed)
-            .unwrap_or(false)
-    }
-
-    /// Check if a blockhash has Candidate status in its metadata.
-    pub fn is_candidate(&self, blockhash: &BlockHash) -> bool {
-        self.get_block_metadata(blockhash)
-            .map(|m| m.status == Status::Candidate)
-            .unwrap_or(false)
-    }
-
-    /// Get branch from a blockhash back to the first ancestor in the candidate chain.
-    ///
-    /// Walks backwards through the chain collecting blockhashes until finding
-    /// one that's already a candidate. Returns the branch including the candidate.
-    pub fn get_branch_to_candidates(
-        &self,
-        blockhash: &BlockHash,
-    ) -> Result<Option<VecDeque<BlockHash>>, StoreError> {
-        let mut branch = VecDeque::with_capacity(BRANCH_INITIAL_CAPACITY);
-
-        let mut current = *blockhash;
-        loop {
-            if self.is_candidate(&current) {
-                // Found ancestor in candidate chain, include it and return
-                branch.push_front(current);
-                return Ok(Some(branch));
-            }
-
-            // Get the share to find its parent
-            let Some(share_header) = self.get_share_header(&current)? else {
-                // Share not found, branch doesn't terminate on cadndidate chain, return empty chain
-                return Ok(None);
-            };
-
-            branch.push_front(current);
-            current = share_header.prev_share_blockhash;
-        }
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-    use crate::test_utils::TestShareBlockBuilder;
-    use tempfile::tempdir;
-
-    // ── increment_top_candidate tests ────────────────────────────────
-
-    #[test]
-    fn test_increment_top_candidate_sets_initial_top_from_height() {
-        let temp_dir = tempdir().unwrap();
-        let store = Store::new(temp_dir.path().to_str().unwrap().to_string(), false).unwrap();
-
-        assert!(store.get_top_candidate_height().is_err());
-
-        let mut batch = Store::get_write_batch();
-        let result = store.increment_top_candidate(5, &mut batch);
-        store.commit_batch(batch).unwrap();
-
-        assert_eq!(result.unwrap(), 5);
-        assert_eq!(store.get_top_candidate_height().unwrap(), 5);
-    }
-
-    #[test]
-    fn test_increment_top_candidate_increments_consecutive_height() {
-        let temp_dir = tempdir().unwrap();
-        let store = Store::new(temp_dir.path().to_str().unwrap().to_string(), false).unwrap();
-
-        // Bootstrap top to 3
-        let mut batch = Store::get_write_batch();
-        store.increment_top_candidate(3, &mut batch).unwrap();
-        store.commit_batch(batch).unwrap();
-
-        // Height 4 is exactly 1 more than current top (3)
-        let mut batch = Store::get_write_batch();
-        let result = store.increment_top_candidate(4, &mut batch);
-        store.commit_batch(batch).unwrap();
-
-        assert_eq!(result.unwrap(), 4);
-        assert_eq!(store.get_top_candidate_height().unwrap(), 4);
-    }
-
-    #[test]
-    fn test_increment_top_candidate_errors_on_skipped_height() {
-        let temp_dir = tempdir().unwrap();
-        let store = Store::new(temp_dir.path().to_str().unwrap().to_string(), false).unwrap();
-
-        let mut batch = Store::get_write_batch();
-        store.increment_top_candidate(1, &mut batch).unwrap();
-        store.commit_batch(batch).unwrap();
-
-        // Height 3 skips over 2
-        let mut batch = Store::get_write_batch();
-        let result = store.increment_top_candidate(3, &mut batch);
-
-        assert!(result.is_err());
-        assert_eq!(store.get_top_candidate_height().unwrap(), 1);
-    }
-
-    #[test]
-    fn test_increment_top_candidate_errors_on_same_height() {
-        let temp_dir = tempdir().unwrap();
-        let store = Store::new(temp_dir.path().to_str().unwrap().to_string(), false).unwrap();
-
-        let mut batch = Store::get_write_batch();
-        store.increment_top_candidate(2, &mut batch).unwrap();
-        store.commit_batch(batch).unwrap();
-
-        // 2 - 2 = 0, not 1
-        let mut batch = Store::get_write_batch();
-        let result = store.increment_top_candidate(2, &mut batch);
-
-        assert!(result.is_err());
-        assert_eq!(store.get_top_candidate_height().unwrap(), 2);
-    }
-
-    // ── append_to_candidate tests ─────────────────────────────────────────
-
-    #[test]
-    fn test_append_to_candidate() {
-        let temp_dir = tempdir().unwrap();
-        let store = Store::new(temp_dir.path().to_str().unwrap().to_string(), false).unwrap();
-
-        let share1 = TestShareBlockBuilder::new()
-            .nonce(0xe9695791)
-            .work(1)
-            .build();
-        let share2 = TestShareBlockBuilder::new()
-            .nonce(0xe9695792)
-            .work(2)
-            .build();
-
-        // Make share1 candidate at height 0
-        let mut batch = Store::get_write_batch();
-        let mut metadata1 = store
-            .add_share(&share1, 1, share1.header.get_work(), false, &mut batch)
-            .unwrap();
-        store
-            .append_to_candidates(&share1.block_hash(), 0, &mut metadata1, &mut batch)
-            .unwrap();
-        store.commit_batch(batch).unwrap();
-
-        // Verify we can retrieve it
-        let candidate = store.get_candidate_at_height(0).unwrap();
-        assert_eq!(candidate, share1.block_hash());
-
-        // Make share2 candidate at height 1
-        let mut batch = Store::get_write_batch();
-        let mut metadata2 = store
-            .add_share(&share2, 2, share2.header.get_work(), false, &mut batch)
-            .unwrap();
-        store
-            .append_to_candidates(&share2.block_hash(), 1, &mut metadata2, &mut batch)
-            .unwrap();
-        store.commit_batch(batch).unwrap();
-
-        // Verify both heights
-        assert_eq!(
-            store.get_candidate_at_height(0).unwrap(),
-            share1.block_hash()
-        );
-        assert_eq!(
-            store.get_candidate_at_height(1).unwrap(),
-            share2.block_hash()
-        );
-
-        // Non-existent height should return None
-        assert!(store.get_candidate_at_height(999).is_err());
-
-        // Top candidate height is changed
-        assert_eq!(store.get_top_candidate_height().unwrap(), 1);
-
-        // Top candidate is changed
-        let top = store.get_top_candidate().unwrap();
-        assert_eq!(top.hash, share2.block_hash());
-        assert_eq!(top.height, 1);
-        assert_eq!(top.work, share2.header.get_work());
-    }
-
-    #[test]
-    fn test_make_confirmed() {
-        let temp_dir = tempdir().unwrap();
-        let store = Store::new(temp_dir.path().to_str().unwrap().to_string(), false).unwrap();
-
-        let share1 = TestShareBlockBuilder::new().nonce(0xe9695791).build();
-        let share2 = TestShareBlockBuilder::new().nonce(0xe9695792).build();
-
-        let mut batch = Store::get_write_batch();
-        let mut metadata1 = store
-            .add_share(&share1, 0, share1.header.get_work(), true, &mut batch)
-            .unwrap();
-        store
-            .append_to_confirmed(&share1.block_hash(), 0, &mut metadata1, &mut batch)
-            .unwrap();
-        store.commit_batch(batch).unwrap();
-
-        // Verify we can retrieve it
-        let confirmed = store.get_confirmed_at_height(0).unwrap();
-        assert_eq!(confirmed, share1.block_hash());
-
-        let mut batch = Store::get_write_batch();
-        let mut metadata2 = store
-            .add_share(&share2, 1, share2.header.get_work(), true, &mut batch)
-            .unwrap();
-        store
-            .append_to_confirmed(&share2.block_hash(), 1, &mut metadata2, &mut batch)
-            .unwrap();
-        store.commit_batch(batch).unwrap();
-
-        // Verify both heights
-        assert_eq!(
-            store.get_confirmed_at_height(0).unwrap(),
-            share1.block_hash()
-        );
-        assert_eq!(
-            store.get_confirmed_at_height(1).unwrap(),
-            share2.block_hash()
-        );
-
-        // Non-existent height should return None
-        assert!(store.get_confirmed_at_height(999).is_err());
-
-        // Top confirmed height is changed
-        assert_eq!(store.get_top_confirmed_height().unwrap(), 1);
-
-        // Top confirmed is changed
-        let top_confirmed = store.get_top_confirmed().unwrap();
-        assert_eq!(top_confirmed.hash, share2.block_hash());
-        assert_eq!(top_confirmed.height, 1);
-        assert_eq!(top_confirmed.work, share2.header.get_work());
-    }
-
-    #[test]
-    fn test_append_to_candidate_on_overwrite_previous_should_error() {
-        // Each height should only have one candidate - new candidates replace old ones
-        let temp_dir = tempdir().unwrap();
-        let store = Store::new(temp_dir.path().to_str().unwrap().to_string(), false).unwrap();
-
-        let share1 = TestShareBlockBuilder::new().nonce(0xe9695791).build();
-        let share2 = TestShareBlockBuilder::new().nonce(0xe9695792).build();
-
-        // Add shares first
-        let mut batch = Store::get_write_batch();
-        let mut metadata1 = store
-            .add_share(&share1, 0, share1.header.get_work(), false, &mut batch)
-            .unwrap();
-        let mut metadata2 = store
-            .add_share(&share2, 0, share2.header.get_work(), false, &mut batch)
-            .unwrap();
-        store
-            .append_to_candidates(&share1.block_hash(), 0, &mut metadata1, &mut batch)
-            .unwrap();
-        store.commit_batch(batch).unwrap();
-
-        assert_eq!(
-            store.get_candidate_at_height(0).unwrap(),
-            share1.block_hash()
-        );
-
-        // Make share2 candidate at same height - should error (height mismatch)
-        let mut batch = Store::get_write_batch();
-        let result =
-            store.append_to_candidates(&share2.block_hash(), 0, &mut metadata2, &mut batch);
-        assert!(result.is_err());
-    }
-
-    #[test]
-    fn test_make_confirmed_ignores_same_height() {
-        // Confirming at the same height as current top is ignored
-        let temp_dir = tempdir().unwrap();
-        let store = Store::new(temp_dir.path().to_str().unwrap().to_string(), false).unwrap();
-
-        let share1 = TestShareBlockBuilder::new().nonce(0xe9695791).build();
-        let share2 = TestShareBlockBuilder::new().nonce(0xe9695792).build();
-        let share3 = TestShareBlockBuilder::new().nonce(0xe9695793).build();
-
-        // Add all shares first
-        let mut batch = Store::get_write_batch();
-        let mut metadata1 = store
-            .add_share(&share1, 0, share1.header.get_work(), false, &mut batch)
-            .unwrap();
-        let mut metadata2 = store
-            .add_share(&share2, 1, share2.header.get_work(), false, &mut batch)
-            .unwrap();
-        let mut metadata3 = store
-            .add_share(&share3, 1, share3.header.get_work(), false, &mut batch)
-            .unwrap();
-        store
-            .append_to_confirmed(&share1.block_hash(), 0, &mut metadata1, &mut batch)
-            .unwrap();
-        store.commit_batch(batch).unwrap();
-
-        assert_eq!(
-            store.get_confirmed_at_height(0).unwrap(),
-            share1.block_hash()
-        );
-
-        // Confirm share2
-        let mut batch = Store::get_write_batch();
-        store
-            .append_to_confirmed(&share2.block_hash(), 1, &mut metadata2, &mut batch)
-            .unwrap();
-        store.commit_batch(batch).unwrap();
-
-        // Confirm share3 at same height as share2, should error.
-        let mut batch = Store::get_write_batch();
-        assert!(
-            store
-                .append_to_confirmed(&share3.block_hash(), 1, &mut metadata3, &mut batch)
-                .is_err()
-        );
-    }
-
-    #[test]
-    fn test_candidate_and_confirmed_are_independent() {
-        // Candidate and confirmed should be stored separately at the same height
-        let temp_dir = tempdir().unwrap();
-        let store = Store::new(temp_dir.path().to_str().unwrap().to_string(), false).unwrap();
-
-        let candidate_share = TestShareBlockBuilder::new().nonce(0xe9695791).build();
-        let confirmed_share = TestShareBlockBuilder::new().nonce(0xe9695792).build();
-
-        // Add shares first
-        let mut batch = Store::get_write_batch();
-        let mut candidate_metadata = store
-            .add_share(
-                &candidate_share,
-                0,
-                candidate_share.header.get_work(),
-                false,
-                &mut batch,
-            )
-            .unwrap();
-        let mut confirmed_metadata = store
-            .add_share(
-                &confirmed_share,
-                0,
-                confirmed_share.header.get_work(),
-                false,
-                &mut batch,
-            )
-            .unwrap();
-        store
-            .append_to_candidates(
-                &candidate_share.block_hash(),
-                0,
-                &mut candidate_metadata,
-                &mut batch,
-            )
-            .unwrap();
-        store.commit_batch(batch).unwrap();
-
-        let mut batch = Store::get_write_batch();
-        store
-            .append_to_confirmed(
-                &confirmed_share.block_hash(),
-                0,
-                &mut confirmed_metadata,
-                &mut batch,
-            )
-            .unwrap();
-        store.commit_batch(batch).unwrap();
-
-        // Both should be retrievable independently
-        assert_eq!(
-            store.get_candidate_at_height(0).unwrap(),
-            candidate_share.block_hash()
-        );
-        assert_eq!(
-            store.get_confirmed_at_height(0).unwrap(),
-            confirmed_share.block_hash()
-        );
-    }
-
-    #[test]
-    fn test_is_confirmed_returns_true_when_confirmed() {
-        let temp_dir = tempdir().unwrap();
-        let store = Store::new(temp_dir.path().to_str().unwrap().to_string(), false).unwrap();
-
-        // Create and add a share to the store (this sets up block metadata)
-        let genesis = TestShareBlockBuilder::new().nonce(0xe9695791).build();
-        let mut batch = Store::get_write_batch();
-        store.setup_genesis(&genesis, &mut batch).unwrap();
-        store.commit_batch(batch).unwrap();
-
-        // Add a child share and mark it confirmed
-        let share1 = TestShareBlockBuilder::new()
-            .prev_share_blockhash(genesis.block_hash().to_string())
-            .nonce(0xe9695792)
-            .build();
-        let mut batch = Store::get_write_batch();
-        let mut metadata = store
-            .add_share(&share1, 1, share1.header.get_work(), true, &mut batch)
-            .unwrap();
-        store
-            .append_to_confirmed(&share1.block_hash(), 1, &mut metadata, &mut batch)
-            .unwrap();
-        store.commit_batch(batch).unwrap();
-
-        // is_confirmed should return true
-        assert!(store.is_confirmed(&share1.block_hash()));
-    }
-
-    #[test]
-    fn test_is_confirmed_returns_true_for_genesis_after_setup() {
-        let temp_dir = tempdir().unwrap();
-        let store = Store::new(temp_dir.path().to_str().unwrap().to_string(), false).unwrap();
-
-        // Create and add a share to the store (this sets up block metadata)
-        let genesis = TestShareBlockBuilder::new().nonce(0xe9695791).build();
-        let mut batch = Store::get_write_batch();
-        store.setup_genesis(&genesis, &mut batch).unwrap();
-        store.commit_batch(batch).unwrap();
-
-        // Genesis is marked confirmed at setup
-        assert!(store.is_confirmed(&genesis.block_hash()));
-    }
-
-    #[test]
-    fn test_is_confirmed_returns_false_when_not_confirmed() {
-        let temp_dir = tempdir().unwrap();
-        let store = Store::new(temp_dir.path().to_str().unwrap().to_string(), false).unwrap();
-
-        // Create and add a share to the store (this sets up block metadata)
-        let genesis = TestShareBlockBuilder::new().nonce(0xe9695791).build();
-        let mut batch = Store::get_write_batch();
-        store.setup_genesis(&genesis, &mut batch).unwrap();
-        store.commit_batch(batch).unwrap();
-
-        // Create another share
-        let share2 = TestShareBlockBuilder::new()
-            .prev_share_blockhash(genesis.block_hash().to_string())
-            .nonce(0xe9695792)
-            .build();
-        let mut batch = Store::get_write_batch();
-        store
-            .add_share(&share2, 1, share2.header.get_work(), true, &mut batch)
-            .unwrap();
-        store.commit_batch(batch).unwrap();
-
-        // Don't mark it as confirmed - is_confirmed should return false
-        assert!(!store.is_confirmed(&share2.block_hash()));
-    }
-
-    #[test]
-    fn test_is_confirmed_returns_false_when_different_block_confirmed() {
-        let temp_dir = tempdir().unwrap();
-        let store = Store::new(temp_dir.path().to_str().unwrap().to_string(), false).unwrap();
-
-        // Create genesis and add to store
-        let genesis = TestShareBlockBuilder::new().nonce(0xe9695791).build();
-        let mut batch = Store::get_write_batch();
-        store.setup_genesis(&genesis, &mut batch).unwrap();
-        store.commit_batch(batch).unwrap();
-
-        // Create another share at the same height
-        let share2 = TestShareBlockBuilder::new()
-            .prev_share_blockhash(genesis.block_hash().to_string())
-            .nonce(0xe9695792)
-            .build();
-        let mut batch = Store::get_write_batch();
-        let mut metadata2 = store
-            .add_share(&share2, 1, share2.header.get_work(), true, &mut batch)
-            .unwrap();
-        store.commit_batch(batch).unwrap();
-
-        // Create a third share at the same height as share2
-        let share3 = TestShareBlockBuilder::new()
-            .prev_share_blockhash(genesis.block_hash().to_string())
-            .nonce(0xe9695793)
-            .build();
-        let mut batch = Store::get_write_batch();
-        store
-            .add_share(&share3, 1, share3.header.get_work(), true, &mut batch)
-            .unwrap();
-        store.commit_batch(batch).unwrap();
-
-        // Mark share2 as confirmed at height 1
-        let mut batch = Store::get_write_batch();
-        store
-            .append_to_confirmed(&share2.block_hash(), 1, &mut metadata2, &mut batch)
-            .unwrap();
-        store.commit_batch(batch).unwrap();
-
-        // share2 should be confirmed, share3 should not
-        assert!(store.is_confirmed(&share2.block_hash()));
-        assert!(!store.is_confirmed(&share3.block_hash()));
-    }
-
-    #[test]
-    fn test_is_confirmed_returns_false_when_no_metadata() {
-        let temp_dir = tempdir().unwrap();
-        let store = Store::new(temp_dir.path().to_str().unwrap().to_string(), false).unwrap();
-
-        // Create a blockhash that doesn't exist in the store
-        let fake_blockhash = "ffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffff"
-            .parse::<BlockHash>()
-            .unwrap();
-
-        // is_confirmed should return false because there's no metadata
-        assert!(!store.is_confirmed(&fake_blockhash));
-    }
-
-    #[test]
-    fn test_append_to_candidate_does_not_update_top_when_height_skips() {
-        let temp_dir = tempdir().unwrap();
-        let store = Store::new(temp_dir.path().to_str().unwrap().to_string(), false).unwrap();
-
-        let share0 = TestShareBlockBuilder::new().nonce(0xe9695791).build();
-        let share2 = TestShareBlockBuilder::new().nonce(0xe9695792).build();
-
-        // Add shares first
-        let mut batch = Store::get_write_batch();
-        let mut metadata0 = store
-            .add_share(&share0, 0, share0.header.get_work(), false, &mut batch)
-            .unwrap();
-        let mut metadata2 = store
-            .add_share(&share2, 2, share2.header.get_work(), false, &mut batch)
-            .unwrap();
-        store
-            .append_to_candidates(&share0.block_hash(), 0, &mut metadata0, &mut batch)
-            .unwrap();
-        store.commit_batch(batch).unwrap();
-
-        // Verify top candidate height is 0
-        assert_eq!(store.get_top_candidate_height().unwrap(), 0);
-
-        // Make share2 candidate at height 2 (skipping height 1)
-        let mut batch = Store::get_write_batch();
-        let result =
-            store.append_to_candidates(&share2.block_hash(), 2, &mut metadata2, &mut batch);
-        assert!(result.is_err());
-
-        // Top candidate height should still be 0 (not updated because we skipped height 1)
-        assert_eq!(store.get_top_candidate_height().unwrap(), 0);
-
-        // The candidate at height 2 is not there
-        assert!(store.get_candidate_at_height(2).is_err());
-    }
-
-    #[test]
-    fn test_make_confirmed_does_not_update_top_when_height_skips() {
-        let temp_dir = tempdir().unwrap();
-        let store = Store::new(temp_dir.path().to_str().unwrap().to_string(), false).unwrap();
-
-        let share0 = TestShareBlockBuilder::new().nonce(0xe9695791).build();
-        let share2 = TestShareBlockBuilder::new().nonce(0xe9695792).build();
-
-        // Add shares first
-        let mut batch = Store::get_write_batch();
-        let mut metadata0 = store
-            .add_share(&share0, 0, share0.header.get_work(), false, &mut batch)
-            .unwrap();
-        let mut metadata2 = store
-            .add_share(&share2, 2, share2.header.get_work(), false, &mut batch)
-            .unwrap();
-        store
-            .append_to_confirmed(&share0.block_hash(), 0, &mut metadata0, &mut batch)
-            .unwrap();
-        store.commit_batch(batch).unwrap();
-
-        // Verify top confirmed height is 0
-        assert_eq!(store.get_top_confirmed_height().unwrap(), 0);
-
-        // Make share2 confirmed at height 2 (skipping height 1) should error
-        let mut batch = Store::get_write_batch();
-        assert!(
-            store
-                .append_to_confirmed(&share2.block_hash(), 2, &mut metadata2, &mut batch)
-                .is_err()
-        );
-    }
-
-    // ── get_candidates / get_candidates_chain tests ───────────────────
-
-    #[test]
-    fn test_get_candidates_returns_blockhashes_in_range() {
-        let temp_dir = tempdir().unwrap();
-        let store = Store::new(temp_dir.path().to_str().unwrap().to_string(), false).unwrap();
-
-        let share0 = TestShareBlockBuilder::new().nonce(0xe9695791).build();
-        let share1 = TestShareBlockBuilder::new().nonce(0xe9695792).build();
-        let share2 = TestShareBlockBuilder::new().nonce(0xe9695793).build();
-
-        let mut batch = Store::get_write_batch();
-        let mut m0 = store
-            .add_share(&share0, 0, share0.header.get_work(), false, &mut batch)
-            .unwrap();
-        store
-            .append_to_candidates(&share0.block_hash(), 0, &mut m0, &mut batch)
-            .unwrap();
-        let mut m1 = store
-            .add_share(&share1, 1, share1.header.get_work(), false, &mut batch)
-            .unwrap();
-        store
-            .append_to_candidates(&share1.block_hash(), 1, &mut m1, &mut batch)
-            .unwrap();
-        let mut m2 = store
-            .add_share(&share2, 2, share2.header.get_work(), false, &mut batch)
-            .unwrap();
-        store
-            .append_to_candidates(&share2.block_hash(), 2, &mut m2, &mut batch)
-            .unwrap();
-        store.commit_batch(batch).unwrap();
-
-        // Full range
-        let result = store.get_candidates(0, 2).unwrap();
-        assert_eq!(result.len(), 3);
-        assert_eq!(result[0], (0, share0.block_hash()));
-        assert_eq!(result[1], (1, share1.block_hash()));
-        assert_eq!(result[2], (2, share2.block_hash()));
-
-        // Sub-range
-        let result = store.get_candidates(1, 2).unwrap();
-        assert_eq!(result.len(), 2);
-        assert_eq!(result[0], (1, share1.block_hash()));
-        assert_eq!(result[1], (2, share2.block_hash()));
-
-        // Single height
-        let result = store.get_candidates(1, 1).unwrap();
-        assert_eq!(result.len(), 1);
-        assert_eq!(result[0], (1, share1.block_hash()));
-    }
-
-    #[test]
-    fn test_get_candidates_returns_empty_when_from_greater_than_to() {
-        let temp_dir = tempdir().unwrap();
-        let store = Store::new(temp_dir.path().to_str().unwrap().to_string(), false).unwrap();
-
-        let result = store.get_candidates(5, 3).unwrap();
-        assert!(result.is_empty());
-    }
-
-    #[test]
-    fn test_get_candidates_chain_returns_candidates_from_blockhash_to_top() {
-        let temp_dir = tempdir().unwrap();
-        let store = Store::new(temp_dir.path().to_str().unwrap().to_string(), false).unwrap();
-
-        let genesis = TestShareBlockBuilder::new().nonce(0xe9695791).build();
-        let mut batch = Store::get_write_batch();
-        store.setup_genesis(&genesis, &mut batch).unwrap();
-        store.commit_batch(batch).unwrap();
-
-        let share1 = TestShareBlockBuilder::new()
-            .prev_share_blockhash(genesis.block_hash().to_string())
-            .nonce(0xe9695792)
-            .build();
-        let share2 = TestShareBlockBuilder::new()
-            .prev_share_blockhash(share1.block_hash().to_string())
-            .nonce(0xe9695793)
-            .build();
-
-        let mut batch = Store::get_write_batch();
-        let mut m1 = store
-            .add_share(&share1, 1, share1.header.get_work(), true, &mut batch)
-            .unwrap();
-        store
-            .append_to_candidates(&share1.block_hash(), 1, &mut m1, &mut batch)
-            .unwrap();
-        let mut m2 = store
-            .add_share(&share2, 2, share2.header.get_work(), true, &mut batch)
-            .unwrap();
-        store
-            .append_to_candidates(&share2.block_hash(), 2, &mut m2, &mut batch)
-            .unwrap();
-        store.commit_batch(batch).unwrap();
-
-        let top = store.get_top_candidate().ok();
-        let result = store
-            .get_candidates_chain(&share1.block_hash(), top.as_ref())
-            .unwrap();
-        assert_eq!(result.len(), 2);
-        assert_eq!(result[0], (1, share1.block_hash()));
-        assert_eq!(result[1], (2, share2.block_hash()));
-    }
-
-    #[test]
-    fn test_get_candidates_chain_errors_when_no_top_candidate() {
-        let temp_dir = tempdir().unwrap();
-        let store = Store::new(temp_dir.path().to_str().unwrap().to_string(), false).unwrap();
-
-        let genesis = TestShareBlockBuilder::new().nonce(0xe9695791).build();
-        let mut batch = Store::get_write_batch();
-        store.setup_genesis(&genesis, &mut batch).unwrap();
-        store.commit_batch(batch).unwrap();
-
-        let result = store.get_candidates_chain(&genesis.block_hash(), None);
-        assert!(result.is_err());
-    }
-
-    // ── get_confirmed / get_confirmed_chain tests ───────────────────
-
-    #[test]
-    fn test_get_confirmed_returns_blockhashes_in_range() {
-        let temp_dir = tempdir().unwrap();
-        let store = Store::new(temp_dir.path().to_str().unwrap().to_string(), false).unwrap();
-
-        let share0 = TestShareBlockBuilder::new().nonce(0xe9695791).build();
-        let share1 = TestShareBlockBuilder::new().nonce(0xe9695792).build();
-        let share2 = TestShareBlockBuilder::new().nonce(0xe9695793).build();
-
-        // append_to_confirmed checks top height against DB, so commit between each
-        let mut batch = Store::get_write_batch();
-        let mut m0 = store
-            .add_share(&share0, 0, share0.header.get_work(), false, &mut batch)
-            .unwrap();
-        store
-            .append_to_confirmed(&share0.block_hash(), 0, &mut m0, &mut batch)
-            .unwrap();
-        store.commit_batch(batch).unwrap();
-
-        let mut batch = Store::get_write_batch();
-        let mut m1 = store
-            .add_share(&share1, 1, share1.header.get_work(), false, &mut batch)
-            .unwrap();
-        store
-            .append_to_confirmed(&share1.block_hash(), 1, &mut m1, &mut batch)
-            .unwrap();
-        store.commit_batch(batch).unwrap();
-
-        let mut batch = Store::get_write_batch();
-        let mut m2 = store
-            .add_share(&share2, 2, share2.header.get_work(), false, &mut batch)
-            .unwrap();
-        store
-            .append_to_confirmed(&share2.block_hash(), 2, &mut m2, &mut batch)
-            .unwrap();
-        store.commit_batch(batch).unwrap();
-
-        // Full range
-        let result = store.get_confirmed(0, 2).unwrap();
-        assert_eq!(result.len(), 3);
-        assert_eq!(result[0], (0, share0.block_hash()));
-        assert_eq!(result[1], (1, share1.block_hash()));
-        assert_eq!(result[2], (2, share2.block_hash()));
-
-        // Sub-range
-        let result = store.get_confirmed(1, 2).unwrap();
-        assert_eq!(result.len(), 2);
-        assert_eq!(result[0], (1, share1.block_hash()));
-        assert_eq!(result[1], (2, share2.block_hash()));
-
-        // Single height
-        let result = store.get_confirmed(1, 1).unwrap();
-        assert_eq!(result.len(), 1);
-        assert_eq!(result[0], (1, share1.block_hash()));
-    }
-
-    #[test]
-    fn test_get_confirmed_returns_empty_when_from_greater_than_to() {
-        let temp_dir = tempdir().unwrap();
-        let store = Store::new(temp_dir.path().to_str().unwrap().to_string(), false).unwrap();
-
-        let result = store.get_confirmed(5, 3).unwrap();
-        assert!(result.is_empty());
-    }
-
-    #[test]
-    fn test_get_confirmed_chain_returns_confirmed_from_blockhash_to_top() {
-        let temp_dir = tempdir().unwrap();
-        let store = Store::new(temp_dir.path().to_str().unwrap().to_string(), false).unwrap();
-
-        let genesis = TestShareBlockBuilder::new().nonce(0xe9695791).build();
-        let mut batch = Store::get_write_batch();
-        store.setup_genesis(&genesis, &mut batch).unwrap();
-        store.commit_batch(batch).unwrap();
-
-        let share1 = TestShareBlockBuilder::new()
-            .prev_share_blockhash(genesis.block_hash().to_string())
-            .nonce(0xe9695792)
-            .build();
-        let share2 = TestShareBlockBuilder::new()
-            .prev_share_blockhash(share1.block_hash().to_string())
-            .nonce(0xe9695793)
-            .build();
-
-        // append_to_confirmed checks top height against DB, so commit between each
-        let mut batch = Store::get_write_batch();
-        let mut m1 = store
-            .add_share(&share1, 1, share1.header.get_work(), true, &mut batch)
-            .unwrap();
-        store
-            .append_to_confirmed(&share1.block_hash(), 1, &mut m1, &mut batch)
-            .unwrap();
-        store.commit_batch(batch).unwrap();
-
-        let mut batch = Store::get_write_batch();
-        let mut m2 = store
-            .add_share(&share2, 2, share2.header.get_work(), true, &mut batch)
-            .unwrap();
-        store
-            .append_to_confirmed(&share2.block_hash(), 2, &mut m2, &mut batch)
-            .unwrap();
-        store.commit_batch(batch).unwrap();
-
-        let top = store.get_top_confirmed().ok();
-        let result = store
-            .get_confirmed_chain(&share1.block_hash(), top.as_ref())
-            .unwrap();
-        assert_eq!(result.len(), 2);
-        assert_eq!(result[0], (1, share1.block_hash()));
-        assert_eq!(result[1], (2, share2.block_hash()));
-    }
-
-    #[test]
-    fn test_get_confirmed_chain_errors_when_no_top_confirmed() {
-        let temp_dir = tempdir().unwrap();
-        let store = Store::new(temp_dir.path().to_str().unwrap().to_string(), false).unwrap();
-
-        let genesis = TestShareBlockBuilder::new().nonce(0xe9695791).build();
-        let mut batch = Store::get_write_batch();
-        store.setup_genesis(&genesis, &mut batch).unwrap();
-        store.commit_batch(batch).unwrap();
-
-        let result = store.get_confirmed_chain(&genesis.block_hash(), None);
-        assert!(result.is_err());
-    }
-
-    // ── is_candidate tests ────────────────────────────────────────────
-
-    #[test]
-    fn test_is_candidate_returns_true_when_candidate() {
-        let temp_dir = tempdir().unwrap();
-        let store = Store::new(temp_dir.path().to_str().unwrap().to_string(), false).unwrap();
-
-        let genesis = TestShareBlockBuilder::new().nonce(0xe9695791).build();
-        let mut batch = Store::get_write_batch();
-        store.setup_genesis(&genesis, &mut batch).unwrap();
-        store.commit_batch(batch).unwrap();
-
-        // Add share and make it a candidate
-        let share = TestShareBlockBuilder::new()
-            .prev_share_blockhash(genesis.block_hash().to_string())
-            .nonce(0xe9695792)
-            .build();
-        let mut batch = Store::get_write_batch();
-        let mut metadata = store
-            .add_share(&share, 1, share.header.get_work(), true, &mut batch)
-            .unwrap();
-        store
-            .append_to_candidates(&share.block_hash(), 1, &mut metadata, &mut batch)
-            .unwrap();
-        store.commit_batch(batch).unwrap();
-
-        assert!(store.is_candidate(&share.block_hash()));
-    }
-
-    #[test]
-    fn test_is_candidate_returns_false_when_not_candidate() {
-        let temp_dir = tempdir().unwrap();
-        let store = Store::new(temp_dir.path().to_str().unwrap().to_string(), false).unwrap();
-
-        let genesis = TestShareBlockBuilder::new().nonce(0xe9695791).build();
-        let mut batch = Store::get_write_batch();
-        store.setup_genesis(&genesis, &mut batch).unwrap();
-        store.commit_batch(batch).unwrap();
-
-        // Add share but don't make it a candidate
-        let share = TestShareBlockBuilder::new()
-            .prev_share_blockhash(genesis.block_hash().to_string())
-            .nonce(0xe9695792)
-            .build();
-        let mut batch = Store::get_write_batch();
-        store
-            .add_share(&share, 1, share.header.get_work(), true, &mut batch)
-            .unwrap();
-        store.commit_batch(batch).unwrap();
-
-        assert!(!store.is_candidate(&share.block_hash()));
-    }
-
-    // ── get_branch_to_candidates tests ────────────────────────────────
-
-    #[test]
-    fn test_get_branch_to_candidates_returns_empty_when_already_candidate() {
-        let temp_dir = tempdir().unwrap();
-        let store = Store::new(temp_dir.path().to_str().unwrap().to_string(), false).unwrap();
-
-        let genesis = TestShareBlockBuilder::new().nonce(0xe9695791).build();
-        let mut batch = Store::get_write_batch();
-        store.setup_genesis(&genesis, &mut batch).unwrap();
-        store.commit_batch(batch).unwrap();
-
-        // Add share and make it a candidate
-        let share = TestShareBlockBuilder::new()
-            .prev_share_blockhash(genesis.block_hash().to_string())
-            .nonce(0xe9695792)
-            .build();
-        let mut batch = Store::get_write_batch();
-        let mut metadata = store
-            .add_share(&share, 1, share.header.get_work(), true, &mut batch)
-            .unwrap();
-        store
-            .append_to_candidates(&share.block_hash(), 1, &mut metadata, &mut batch)
-            .unwrap();
-        store.commit_batch(batch).unwrap();
-
-        // Branch should be empty since share is already a candidate
-        let branch = store.get_branch_to_candidates(&share.block_hash()).unwrap();
-        assert!(branch.is_some());
-        assert_eq!(branch, Some(VecDeque::from([share.block_hash()])));
-    }
-
-    #[test]
-    fn test_get_branch_to_candidates_returns_branch_to_candidate_ancestor() {
-        let temp_dir = tempdir().unwrap();
-        let store = Store::new(temp_dir.path().to_str().unwrap().to_string(), false).unwrap();
-
-        // Setup: genesis -> share1 (candidate) -> share2 -> share3
-        let genesis = TestShareBlockBuilder::new().nonce(0xe9695791).build();
-        let mut batch = Store::get_write_batch();
-        store.setup_genesis(&genesis, &mut batch).unwrap();
-        store.commit_batch(batch).unwrap();
-
-        // share1 is a candidate at height 1
-        let share1 = TestShareBlockBuilder::new()
-            .prev_share_blockhash(genesis.block_hash().to_string())
-            .nonce(0xe9695792)
-            .build();
-        let mut batch = Store::get_write_batch();
-        let mut metadata1 = store
-            .add_share(&share1, 1, share1.header.get_work(), true, &mut batch)
-            .unwrap();
-        store
-            .append_to_candidates(&share1.block_hash(), 1, &mut metadata1, &mut batch)
-            .unwrap();
-        store.commit_batch(batch).unwrap();
-
-        // share2 extends share1 but is NOT on candidate chain
-        let share2 = TestShareBlockBuilder::new()
-            .prev_share_blockhash(share1.block_hash().to_string())
-            .nonce(0xe9695793)
-            .build();
-        let mut batch = Store::get_write_batch();
-        store
-            .add_share(&share2, 2, share2.header.get_work(), true, &mut batch)
-            .unwrap();
-        store.commit_batch(batch).unwrap();
-
-        // share3 extends share2 and is NOT on candidate chain
-        let share3 = TestShareBlockBuilder::new()
-            .prev_share_blockhash(share2.block_hash().to_string())
-            .nonce(0xe9695794)
-            .build();
-        let mut batch = Store::get_write_batch();
-        store
-            .add_share(&share3, 3, share3.header.get_work(), true, &mut batch)
-            .unwrap();
-        store.commit_batch(batch).unwrap();
-
-        // Branch from share3 should be [share1, share2, share3]
-        let branch = store
-            .get_branch_to_candidates(&share3.block_hash())
-            .unwrap()
-            .unwrap();
-        assert_eq!(branch.len(), 3);
-        assert_eq!(branch[0], share1.block_hash());
-        assert_eq!(branch[1], share2.block_hash());
-        assert_eq!(branch[2], share3.block_hash());
-    }
-
-    #[test]
-    fn test_get_branch_to_candidates_returns_single_share_branch() {
-        let temp_dir = tempdir().unwrap();
-        let store = Store::new(temp_dir.path().to_str().unwrap().to_string(), false).unwrap();
-
-        // Setup: genesis -> share1 (candidate) -> share2 (not on candidate chain)
-        let genesis = TestShareBlockBuilder::new().nonce(0xe9695791).build();
-        let mut batch = Store::get_write_batch();
-        store.setup_genesis(&genesis, &mut batch).unwrap();
-        store.commit_batch(batch).unwrap();
-
-        let share1 = TestShareBlockBuilder::new()
-            .prev_share_blockhash(genesis.block_hash().to_string())
-            .nonce(0xe9695792)
-            .build();
-        let mut batch = Store::get_write_batch();
-        let mut metadata1 = store
-            .add_share(&share1, 1, share1.header.get_work(), true, &mut batch)
-            .unwrap();
-        store
-            .append_to_candidates(&share1.block_hash(), 1, &mut metadata1, &mut batch)
-            .unwrap();
-        store.commit_batch(batch).unwrap();
-
-        let share2 = TestShareBlockBuilder::new()
-            .prev_share_blockhash(share1.block_hash().to_string())
-            .nonce(0xe9695793)
-            .build();
-        let mut batch = Store::get_write_batch();
-        store
-            .add_share(&share2, 2, share2.header.get_work(), true, &mut batch)
-            .unwrap();
-        store.commit_batch(batch).unwrap();
-
-        // Branch from share2 should be just [share1, share2]
-        let branch = store
-            .get_branch_to_candidates(&share2.block_hash())
-            .unwrap()
-            .unwrap();
-        assert_eq!(branch.len(), 2);
-        assert_eq!(branch[0], share1.block_hash());
-        assert_eq!(branch[1], share2.block_hash());
     }
 }
