+++ p2poolv2_lib/src/store/organise/organise_share.rs
@@ -148,264 +148,6 @@ impl Store {
 
         Ok(None)
     }
-
-    /// Check if the confirmed chain can be extended by the local candidate chain.
-    ///
-    /// Accepts the effective candidate chain built locally (avoiding stale
-    /// reads from the DB within the same WriteBatch).
-    /// Returns true if the first candidate is a child of the top confirmed.
-    fn should_extend_confirmed(
-        &self,
-        candidates: &Chain,
-        top_confirmed_height: Height,
-        top_confirmed_hash: BlockHash,
-    ) -> Result<bool, StoreError> {
-        debug!(
-            "top confirmed height {}, top confirmed hash {}",
-            top_confirmed_height, top_confirmed_hash
-        );
-
-        if candidates.is_empty() {
-            debug!("No candidates found");
-            return Ok(false);
-        }
-
-        let Some(first_candidate_header) = self.get_share_header(&candidates[0].1)? else {
-            return Err(StoreError::NotFound(
-                "No candidate header found in extending confirmed".into(),
-            ));
-        };
-
-        debug!("First candidate header {:?}", first_candidate_header);
-
-        // First candidate must be a child of top confirmed
-        Ok(
-            first_candidate_header.prev_share_blockhash == top_confirmed_hash
-                && top_confirmed_height + 1 == candidates[0].0,
-        )
-    }
-
-    /// Promote candidates to confirmed and clear the candidate chain.
-    ///
-    /// Moves each candidate entry to the confirmed index, updates metadata
-    /// to Confirmed status, and removes the top candidate height marker.
-    fn extend_confirmed(
-        &self,
-        to: Height,
-        candidates: Vec<(Height, BlockHash)>,
-        batch: &mut rocksdb::WriteBatch,
-    ) -> Result<Option<Height>, StoreError> {
-        for (candidate_height, candidate_hash) in candidates {
-            self.put_confirmed_entry(candidate_height, &candidate_hash, batch);
-            self.delete_candidate_entry(candidate_height, batch);
-            let mut metadata = self.get_block_metadata(&candidate_hash)?;
-            metadata.status = crate::store::block_tx_metadata::Status::Confirmed;
-            self.update_block_metadata(&candidate_hash, &metadata, batch)?;
-        }
-        self.delete_top_candidate_height(batch);
-        self.set_top_confirmed_height(to, batch);
-        Ok(Some(to))
-    }
-
-    /// Returns true if the share being organised has more cumulative
-    /// work than the top candidate. This identifies the case where
-    /// share is not building on the current top_candidate, but is a
-    /// different branch that needs to be reorged in.
-    fn should_reorg_candidate(
-        &self,
-        share_blockhash: &BlockHash,
-        metadata: &BlockMetadata,
-        top_candidate: Option<&TopResult>,
-    ) -> bool {
-        match top_candidate {
-            Some(top) => metadata.chain_work > top.work && top.hash != *share_blockhash,
-            None => false,
-        }
-    }
-
-    /// Reorgs the candidate chain to the branch ending at `blockhash`.
-    ///
-    /// Returns the new top candidate height and the new candidate chain
-    /// as `(height, blockhash)` pairs. The caller uses this local chain
-    /// to check confirmed extension without re-reading from the DB.
-    ///
-    /// Directly manipulates candidate index entries and sets the final
-    /// top height in a single pass, avoiding stale reads from the DB
-    /// within the same WriteBatch. Reorged-out shares have their
-    /// metadata status set to Valid so `is_candidate()` stays correct.
-    fn reorg_candidate(
-        &self,
-        blockhash: &BlockHash,
-        top_candidate: Option<&TopResult>,
-        batch: &mut rocksdb::WriteBatch,
-    ) -> Result<(Height, Chain), StoreError> {
-        let branch = self.get_branch_to_candidates(blockhash)?.ok_or_else(|| {
-            StoreError::NotFound("Branch point to reorg candidate chain not found.".into())
-        })?;
-        let branch_point = branch.front().ok_or_else(|| {
-            StoreError::NotFound("Empty branch returned from get_branch_to_candidates.".into())
-        })?;
-        let reorged_out_chain = self.get_candidates_chain(branch_point, top_candidate)?;
-
-        // Collect new branch blockhashes for quick lookup when deciding
-        // whether to set reorged-out metadata to Valid
-        let branch_set: std::collections::HashSet<&BlockHash> = branch.iter().collect();
-
-        // Delete old candidate index entries and set reorged-out shares to Valid
-        for (height, uncandidate) in &reorged_out_chain {
-            self.delete_candidate_entry(*height, batch);
-            if !branch_set.contains(uncandidate) {
-                let mut metadata = self.get_block_metadata(uncandidate)?;
-                metadata.status = crate::store::block_tx_metadata::Status::Valid;
-                self.update_block_metadata(uncandidate, &metadata, batch)?;
-            }
-        }
-
-        // Write new branch entries, collect the chain, and update metadata
-        let mut new_top_height = 0u32;
-        let mut new_chain = Vec::with_capacity(branch.len());
-        for candidate in &branch {
-            let mut metadata = self.get_block_metadata(candidate)?;
-            let height = metadata.expected_height.ok_or_else(|| {
-                StoreError::NotFound("Block metadata missing expected_height for candidate".into())
-            })?;
-            self.put_candidate_entry(height, candidate, batch);
-
-            if metadata.status != crate::store::block_tx_metadata::Status::Candidate {
-                metadata.status = crate::store::block_tx_metadata::Status::Candidate;
-                self.update_block_metadata(candidate, &metadata, batch)?;
-            }
-
-            new_chain.push((height, *candidate));
-            new_top_height = height;
-        }
-
-        // Set the final top candidate height directly
-        self.set_top_candidate_height(new_top_height, batch);
-        Ok((new_top_height, new_chain))
-    }
-
-    /// Walk forward from the current candidate tip, discovering children
-    /// already stored with `Valid` status and appending them to the
-    /// candidate chain in the same WriteBatch.
-    ///
-    /// Among multiple children at the same height, selects the one with
-    /// the highest `chain_work`. Verifies parent hash to exclude uncle
-    /// relationships from the block index. Overwrites `top_candidate_height`
-    /// only if at least one child was appended.
-    fn extend_candidates_with_children(
-        &self,
-        current_top_height: Height,
-        current_top_hash: &BlockHash,
-        candidates: &mut Chain,
-        batch: &mut rocksdb::WriteBatch,
-    ) -> Result<Height, StoreError> {
-        let mut height = current_top_height;
-        let mut tip_hash = *current_top_hash;
-        let mut found_child = true;
-
-        while found_child {
-            found_child = false;
-
-            let children = self
-                .get_children_blockhashes(&tip_hash)?
-                .unwrap_or_default();
-
-            if let Some((best_hash, mut best_metadata)) =
-                self.pick_best_child(&children, &tip_hash, height + 1)?
-            {
-                let next_height = height + 1;
-                self.put_candidate_entry(next_height, &best_hash, batch);
-                best_metadata.status = crate::store::block_tx_metadata::Status::Candidate;
-                self.update_block_metadata(&best_hash, &best_metadata, batch)?;
-                candidates.push((next_height, best_hash));
-
-                height = next_height;
-                tip_hash = best_hash;
-                found_child = true;
-            }
-        }
-
-        if height > current_top_height {
-            self.set_top_candidate_height(height, batch);
-        }
-        Ok(height)
-    }
-
-    /// Select the best qualifying child from a list of children.
-    ///
-    /// Filters for `Valid` status, correct `expected_height`, and matching
-    /// parent hash (to exclude uncle links in the block index). Among
-    /// qualifying children, returns the one with the highest `chain_work`.
-    fn pick_best_child(
-        &self,
-        children: &[BlockHash],
-        parent_hash: &BlockHash,
-        expected_height: Height,
-    ) -> Result<Option<(BlockHash, BlockMetadata)>, StoreError> {
-        let mut top_work_child: Option<(BlockHash, BlockMetadata)> = None;
-
-        for child_hash in children {
-            let all_children = self
-                .get_block_metadata(child_hash)
-                .ok()
-                .filter(|m| m.status == crate::store::block_tx_metadata::Status::Valid)
-                .filter(|m| m.expected_height == Some(expected_height))
-                .and_then(|m| {
-                    self.get_share_header(child_hash)
-                        .ok()
-                        .flatten()
-                        .filter(|h| h.prev_share_blockhash == *parent_hash)
-                        .map(|_| (*child_hash, m))
-                });
-
-            if let Some((child_hash, child_metadata)) = all_children {
-                let has_more_work = top_work_child
-                    .as_ref()
-                    .map(|(_, current_top_metadata)| {
-                        child_metadata.chain_work > current_top_metadata.chain_work
-                    })
-                    .unwrap_or(true);
-                if has_more_work {
-                    top_work_child = Some((child_hash, child_metadata));
-                }
-            }
-        }
-
-        Ok(top_work_child)
-    }
-
-    /// Extends candidate chain, if:
-    /// 1. new share's height is one more than top candidate
-    /// 2. new share's prev hash is top candidate hash
-    /// 3. new share's chain work is more than top candidate's chain work.
-    /// 4. Or, adds to candidate chain if it is empty.
-    ///
-    /// The same function is used to check if the share extending the
-    /// top_candidate or top_confirmed chains using `top_at_chain` param.
-    ///
-    /// Returns true if candidate chain is extended.
-    fn should_extend_candidates(
-        &self,
-        share: &ShareBlock,
-        metadata: &BlockMetadata,
-        top_at_chain: Option<&TopResult>,
-    ) -> Result<Option<Height>, StoreError> {
-        match top_at_chain {
-            None => Ok(metadata.expected_height),
-            Some(top) => {
-                let expected_height = metadata.expected_height.unwrap_or_default();
-                if top.hash == share.header.prev_share_blockhash
-                    && expected_height == top.height + 1
-                    && metadata.chain_work > top.work
-                {
-                    Ok(Some(expected_height))
-                } else {
-                    Ok(None)
-                }
-            }
-        }
-    }
 }
 
 #[cfg(test)]
@@ -414,138 +156,7 @@ mod tests {
     use crate::store::block_tx_metadata::Status;
     use crate::store::organise::TopResult;
     use crate::test_utils::TestShareBlockBuilder;
-    use bitcoin::Work;
     use tempfile::tempdir;
-
-    // ── extend_candidates_at unit tests ──────────────────────────────
-
-    #[test]
-    fn test_extend_candidates_at_returns_share_expected_height_when_no_top_candidate() {
-        let temp_dir = tempdir().unwrap();
-        let store = Store::new(temp_dir.path().to_str().unwrap().to_string(), false).unwrap();
-
-        let share = TestShareBlockBuilder::new().nonce(0xe9695791).build();
-        let metadata = BlockMetadata {
-            expected_height: Some(5),
-            chain_work: share.header.get_work(),
-            status: Status::Pending,
-        };
-
-        let result = store.should_extend_candidates(&share, &metadata, None);
-        assert_eq!(result.unwrap(), metadata.expected_height);
-    }
-
-    #[test]
-    fn test_extend_candidates_at_returns_height_when_conditions_match() {
-        let temp_dir = tempdir().unwrap();
-        let store = Store::new(temp_dir.path().to_str().unwrap().to_string(), false).unwrap();
-
-        let parent = TestShareBlockBuilder::new().nonce(0xe9695791).build();
-        let parent_hash = parent.block_hash();
-
-        let share = TestShareBlockBuilder::new()
-            .prev_share_blockhash(parent_hash.to_string())
-            .nonce(0xe9695792)
-            .build();
-
-        let metadata = BlockMetadata {
-            expected_height: Some(6),
-            chain_work: share.header.get_work(),
-            status: Status::Pending,
-        };
-
-        // height == top candidate height + 1 → 6 == 5 + 1
-        let top_candidate = Some(TopResult {
-            hash: parent_hash,
-            height: 5,
-            work: Work::from_hex("0x05").unwrap(),
-        });
-
-        let result = store.should_extend_candidates(&share, &metadata, top_candidate.as_ref());
-        assert_eq!(result.unwrap(), Some(6));
-    }
-
-    #[test]
-    fn test_extend_candidates_at_returns_none_when_hash_mismatch() {
-        let temp_dir = tempdir().unwrap();
-        let store = Store::new(temp_dir.path().to_str().unwrap().to_string(), false).unwrap();
-
-        let share = TestShareBlockBuilder::new().nonce(0xe9695791).build();
-        let different_share = TestShareBlockBuilder::new().nonce(0xe9695792).build();
-        let different_hash = different_share.block_hash();
-
-        let metadata = BlockMetadata {
-            expected_height: Some(6),
-            chain_work: share.header.get_work(),
-            status: Status::Pending,
-        };
-
-        // Height condition met (6 == 5+1), but hash differs from prev_share_blockhash
-        let top_candidate = Some(TopResult {
-            hash: different_hash,
-            height: 5,
-            work: Work::from_hex("0x05").unwrap(),
-        });
-
-        let result = store.should_extend_candidates(&share, &metadata, top_candidate.as_ref());
-        assert_eq!(result.unwrap(), None);
-    }
-
-    #[test]
-    fn test_extend_candidates_at_returns_none_when_height_mismatch() {
-        let temp_dir = tempdir().unwrap();
-        let store = Store::new(temp_dir.path().to_str().unwrap().to_string(), false).unwrap();
-
-        let parent = TestShareBlockBuilder::new().nonce(0xe9695791).build();
-        let parent_hash = parent.block_hash();
-
-        let share = TestShareBlockBuilder::new()
-            .prev_share_blockhash(parent_hash.to_string())
-            .nonce(0xe9695792)
-            .build();
-
-        let metadata = BlockMetadata {
-            expected_height: Some(7),
-            chain_work: share.header.get_work(),
-            status: Status::Pending,
-        };
-
-        // Hash matches but height doesn't (7 != 5+1)
-        let top_candidate = Some(TopResult {
-            hash: parent_hash,
-            height: 5,
-            work: Work::from_hex("0x05").unwrap(),
-        });
-
-        let result = store.should_extend_candidates(&share, &metadata, top_candidate.as_ref());
-        assert_eq!(result.unwrap(), None);
-    }
-
-    #[test]
-    fn test_extend_candidates_at_returns_none_when_both_mismatch() {
-        let temp_dir = tempdir().unwrap();
-        let store = Store::new(temp_dir.path().to_str().unwrap().to_string(), false).unwrap();
-
-        let share = TestShareBlockBuilder::new().nonce(0xe9695791).build();
-        let different_share = TestShareBlockBuilder::new().nonce(0xe9695792).build();
-        let different_hash = different_share.block_hash();
-
-        let metadata = BlockMetadata {
-            expected_height: Some(5),
-            chain_work: share.header.get_work(),
-            status: Status::Pending,
-        };
-
-        // Neither hash nor height matches
-        let top_candidate = Some(TopResult {
-            hash: different_hash,
-            height: 10,
-            work: Work::from_hex("0x05").unwrap(),
-        });
-
-        let result = store.should_extend_candidates(&share, &metadata, top_candidate.as_ref());
-        assert_eq!(result.unwrap(), None);
-    }
 
     // ── organise_share integration tests ─────────────────────────────
 
@@ -710,132 +321,6 @@ mod tests {
                 work: genesis.header.get_work()
             })
         );
-    }
-
-    // ── should_reorg_candidate unit tests ──────────────────────────────
-
-    #[test]
-    fn test_should_reorg_candidate_true_when_more_work_different_hash() {
-        let temp_dir = tempdir().unwrap();
-        let store = Store::new(temp_dir.path().to_str().unwrap().to_string(), false).unwrap();
-
-        let share = TestShareBlockBuilder::new().nonce(0xe9695791).build();
-        let top_share = TestShareBlockBuilder::new().nonce(0xe9695792).build();
-
-        let metadata = BlockMetadata {
-            expected_height: Some(2),
-            chain_work: Work::from_hex("0x10").unwrap(),
-            status: Status::Valid,
-        };
-
-        let top_candidate = Some(TopResult {
-            hash: top_share.block_hash(),
-            height: 2,
-            work: Work::from_hex("0x05").unwrap(),
-        });
-
-        assert!(store.should_reorg_candidate(
-            &share.block_hash(),
-            &metadata,
-            top_candidate.as_ref()
-        ));
-    }
-
-    #[test]
-    fn test_should_reorg_candidate_false_when_less_work() {
-        let temp_dir = tempdir().unwrap();
-        let store = Store::new(temp_dir.path().to_str().unwrap().to_string(), false).unwrap();
-
-        let share = TestShareBlockBuilder::new().nonce(0xe9695791).build();
-        let top_share = TestShareBlockBuilder::new().nonce(0xe9695792).build();
-
-        let metadata = BlockMetadata {
-            expected_height: Some(2),
-            chain_work: Work::from_hex("0x03").unwrap(),
-            status: Status::Valid,
-        };
-
-        let top_candidate = Some(TopResult {
-            hash: top_share.block_hash(),
-            height: 2,
-            work: Work::from_hex("0x05").unwrap(),
-        });
-
-        assert!(!store.should_reorg_candidate(
-            &share.block_hash(),
-            &metadata,
-            top_candidate.as_ref()
-        ));
-    }
-
-    #[test]
-    fn test_should_reorg_candidate_false_when_equal_work() {
-        let temp_dir = tempdir().unwrap();
-        let store = Store::new(temp_dir.path().to_str().unwrap().to_string(), false).unwrap();
-
-        let share = TestShareBlockBuilder::new().nonce(0xe9695791).build();
-        let top_share = TestShareBlockBuilder::new().nonce(0xe9695792).build();
-
-        let metadata = BlockMetadata {
-            expected_height: Some(2),
-            chain_work: Work::from_hex("0x05").unwrap(),
-            status: Status::Valid,
-        };
-
-        let top_candidate = Some(TopResult {
-            hash: top_share.block_hash(),
-            height: 2,
-            work: Work::from_hex("0x05").unwrap(),
-        });
-
-        assert!(!store.should_reorg_candidate(
-            &share.block_hash(),
-            &metadata,
-            top_candidate.as_ref()
-        ));
-    }
-
-    #[test]
-    fn test_should_reorg_candidate_false_when_same_hash() {
-        let temp_dir = tempdir().unwrap();
-        let store = Store::new(temp_dir.path().to_str().unwrap().to_string(), false).unwrap();
-
-        let share = TestShareBlockBuilder::new().nonce(0xe9695791).build();
-
-        let metadata = BlockMetadata {
-            expected_height: Some(2),
-            chain_work: Work::from_hex("0x10").unwrap(),
-            status: Status::Candidate,
-        };
-
-        // Same blockhash as share — should not reorg against itself
-        let top_candidate = Some(TopResult {
-            hash: share.block_hash(),
-            height: 2,
-            work: Work::from_hex("0x05").unwrap(),
-        });
-
-        assert!(!store.should_reorg_candidate(
-            &share.block_hash(),
-            &metadata,
-            top_candidate.as_ref()
-        ));
-    }
-
-    #[test]
-    fn test_should_reorg_candidate_false_when_no_top_candidate() {
-        let temp_dir = tempdir().unwrap();
-        let store = Store::new(temp_dir.path().to_str().unwrap().to_string(), false).unwrap();
-
-        let share = TestShareBlockBuilder::new().nonce(0xe9695791).build();
-
-        let metadata = BlockMetadata {
-            expected_height: Some(2),
-            chain_work: Work::from_hex("0x10").unwrap(),
-            status: Status::Valid,
-        };
-
-        assert!(!store.should_reorg_candidate(&share.block_hash(), &metadata, None));
     }
 
     // ── reorg_candidate integration tests ──────────────────────────────
@@ -1600,265 +1085,5 @@ mod tests {
         );
         assert_eq!(store.get_top_confirmed_height().unwrap(), 2);
         assert!(store.get_confirmed_at_height(3).is_err());
-    }
-
-    // ── pick_best_child unit tests ──────────────────────────────────
-
-    #[test]
-    fn test_pick_best_child_returns_none_for_empty_children() {
-        let temp_dir = tempdir().unwrap();
-        let store = Store::new(temp_dir.path().to_str().unwrap().to_string(), false).unwrap();
-
-        let genesis = TestShareBlockBuilder::new().nonce(0xe9695791).build();
-        let result = store
-            .pick_best_child(&[], &genesis.block_hash(), 1)
-            .unwrap();
-        assert!(result.is_none());
-    }
-
-    #[test]
-    fn test_pick_best_child_returns_none_when_no_metadata() {
-        let temp_dir = tempdir().unwrap();
-        let store = Store::new(temp_dir.path().to_str().unwrap().to_string(), false).unwrap();
-
-        // A blockhash that was never added to the store
-        let fake_hash = TestShareBlockBuilder::new()
-            .nonce(0xe9695799)
-            .build()
-            .block_hash();
-        let parent = TestShareBlockBuilder::new().nonce(0xe9695791).build();
-
-        let result = store
-            .pick_best_child(&[fake_hash], &parent.block_hash(), 1)
-            .unwrap();
-        assert!(result.is_none());
-    }
-
-    #[test]
-    fn test_pick_best_child_skips_candidate_status() {
-        let temp_dir = tempdir().unwrap();
-        let store = Store::new(temp_dir.path().to_str().unwrap().to_string(), false).unwrap();
-
-        let genesis = TestShareBlockBuilder::new().nonce(0xe9695791).build();
-        let mut batch = Store::get_write_batch();
-        store.setup_genesis(&genesis, &mut batch).unwrap();
-        store.commit_batch(batch).unwrap();
-
-        let child = TestShareBlockBuilder::new()
-            .prev_share_blockhash(genesis.block_hash().to_string())
-            .nonce(0xe9695792)
-            .build();
-        let mut batch = Store::get_write_batch();
-        let mut m = store
-            .add_share(&child, 1, child.header.get_work(), true, &mut batch)
-            .unwrap();
-        // Mark as Candidate — should be skipped by pick_best_child
-        store
-            .append_to_candidates(&child.block_hash(), 1, &mut m, &mut batch)
-            .unwrap();
-        store.commit_batch(batch).unwrap();
-
-        let result = store
-            .pick_best_child(&[child.block_hash()], &genesis.block_hash(), 1)
-            .unwrap();
-        assert!(result.is_none());
-    }
-
-    #[test]
-    fn test_pick_best_child_skips_wrong_height() {
-        let temp_dir = tempdir().unwrap();
-        let store = Store::new(temp_dir.path().to_str().unwrap().to_string(), false).unwrap();
-
-        let genesis = TestShareBlockBuilder::new().nonce(0xe9695791).build();
-        let mut batch = Store::get_write_batch();
-        store.setup_genesis(&genesis, &mut batch).unwrap();
-        store.commit_batch(batch).unwrap();
-
-        let child = TestShareBlockBuilder::new()
-            .prev_share_blockhash(genesis.block_hash().to_string())
-            .nonce(0xe9695792)
-            .build();
-        let mut batch = Store::get_write_batch();
-        store
-            .add_share(&child, 1, child.header.get_work(), true, &mut batch)
-            .unwrap();
-        store.commit_batch(batch).unwrap();
-
-        // Ask for height 5 — child is at height 1
-        let result = store
-            .pick_best_child(&[child.block_hash()], &genesis.block_hash(), 5)
-            .unwrap();
-        assert!(result.is_none());
-    }
-
-    #[test]
-    fn test_pick_best_child_skips_wrong_parent() {
-        let temp_dir = tempdir().unwrap();
-        let store = Store::new(temp_dir.path().to_str().unwrap().to_string(), false).unwrap();
-
-        let genesis = TestShareBlockBuilder::new().nonce(0xe9695791).build();
-        let mut batch = Store::get_write_batch();
-        store.setup_genesis(&genesis, &mut batch).unwrap();
-        store.commit_batch(batch).unwrap();
-
-        // child's prev_share_blockhash is genesis, not some_other
-        let child = TestShareBlockBuilder::new()
-            .prev_share_blockhash(genesis.block_hash().to_string())
-            .nonce(0xe9695792)
-            .build();
-        let mut batch = Store::get_write_batch();
-        store
-            .add_share(&child, 1, child.header.get_work(), true, &mut batch)
-            .unwrap();
-        store.commit_batch(batch).unwrap();
-
-        let wrong_parent = TestShareBlockBuilder::new().nonce(0xe9695799).build();
-        let result = store
-            .pick_best_child(&[child.block_hash()], &wrong_parent.block_hash(), 1)
-            .unwrap();
-        assert!(result.is_none());
-    }
-
-    #[test]
-    fn test_pick_best_child_returns_valid_child() {
-        let temp_dir = tempdir().unwrap();
-        let store = Store::new(temp_dir.path().to_str().unwrap().to_string(), false).unwrap();
-
-        let genesis = TestShareBlockBuilder::new().nonce(0xe9695791).build();
-        let mut batch = Store::get_write_batch();
-        store.setup_genesis(&genesis, &mut batch).unwrap();
-        store.commit_batch(batch).unwrap();
-
-        let child = TestShareBlockBuilder::new()
-            .prev_share_blockhash(genesis.block_hash().to_string())
-            .nonce(0xe9695792)
-            .build();
-        let mut batch = Store::get_write_batch();
-        store
-            .add_share(&child, 1, child.header.get_work(), true, &mut batch)
-            .unwrap();
-        store.commit_batch(batch).unwrap();
-
-        let result = store
-            .pick_best_child(&[child.block_hash()], &genesis.block_hash(), 1)
-            .unwrap();
-        assert!(result.is_some());
-        let (hash, metadata) = result.unwrap();
-        assert_eq!(hash, child.block_hash());
-        assert_eq!(metadata.expected_height, Some(1));
-    }
-
-    #[test]
-    fn test_pick_best_child_selects_heaviest_among_multiple() {
-        let temp_dir = tempdir().unwrap();
-        let store = Store::new(temp_dir.path().to_str().unwrap().to_string(), false).unwrap();
-
-        let genesis = TestShareBlockBuilder::new().nonce(0xe9695791).build();
-        let mut batch = Store::get_write_batch();
-        store.setup_genesis(&genesis, &mut batch).unwrap();
-        store.commit_batch(batch).unwrap();
-
-        let light = TestShareBlockBuilder::new()
-            .prev_share_blockhash(genesis.block_hash().to_string())
-            .work(1)
-            .nonce(0xe9695792)
-            .build();
-        let heavy = TestShareBlockBuilder::new()
-            .prev_share_blockhash(genesis.block_hash().to_string())
-            .work(3)
-            .nonce(0xe9695793)
-            .build();
-
-        let mut batch = Store::get_write_batch();
-        store
-            .add_share(&light, 1, light.header.get_work(), true, &mut batch)
-            .unwrap();
-        store.commit_batch(batch).unwrap();
-
-        let mut batch = Store::get_write_batch();
-        store
-            .add_share(&heavy, 1, heavy.header.get_work(), true, &mut batch)
-            .unwrap();
-        store.commit_batch(batch).unwrap();
-
-        // Both are valid children at height 1 — heavy has more work
-        let result = store
-            .pick_best_child(
-                &[light.block_hash(), heavy.block_hash()],
-                &genesis.block_hash(),
-                1,
-            )
-            .unwrap();
-        assert_eq!(result.unwrap().0, heavy.block_hash());
-
-        // Order shouldn't matter — reverse the input
-        let result = store
-            .pick_best_child(
-                &[heavy.block_hash(), light.block_hash()],
-                &genesis.block_hash(),
-                1,
-            )
-            .unwrap();
-        assert_eq!(result.unwrap().0, heavy.block_hash());
-    }
-
-    #[test]
-    fn test_pick_best_child_skips_invalid_among_valid() {
-        let temp_dir = tempdir().unwrap();
-        let store = Store::new(temp_dir.path().to_str().unwrap().to_string(), false).unwrap();
-
-        let genesis = TestShareBlockBuilder::new().nonce(0xe9695791).build();
-        let mut batch = Store::get_write_batch();
-        store.setup_genesis(&genesis, &mut batch).unwrap();
-        store.commit_batch(batch).unwrap();
-
-        // valid_child: Status::Valid (default from add_share)
-        let valid_child = TestShareBlockBuilder::new()
-            .prev_share_blockhash(genesis.block_hash().to_string())
-            .work(1)
-            .nonce(0xe9695792)
-            .build();
-        let mut batch = Store::get_write_batch();
-        store
-            .add_share(
-                &valid_child,
-                1,
-                valid_child.header.get_work(),
-                true,
-                &mut batch,
-            )
-            .unwrap();
-        store.commit_batch(batch).unwrap();
-
-        // candidate_child: marked Candidate (more work but ineligible)
-        let candidate_child = TestShareBlockBuilder::new()
-            .prev_share_blockhash(genesis.block_hash().to_string())
-            .work(3)
-            .nonce(0xe9695793)
-            .build();
-        let mut batch = Store::get_write_batch();
-        let mut m = store
-            .add_share(
-                &candidate_child,
-                1,
-                candidate_child.header.get_work(),
-                true,
-                &mut batch,
-            )
-            .unwrap();
-        store
-            .append_to_candidates(&candidate_child.block_hash(), 1, &mut m, &mut batch)
-            .unwrap();
-        store.commit_batch(batch).unwrap();
-
-        // candidate_child has more work but is Candidate status — should be skipped
-        let result = store
-            .pick_best_child(
-                &[candidate_child.block_hash(), valid_child.block_hash()],
-                &genesis.block_hash(),
-                1,
-            )
-            .unwrap();
-        assert_eq!(result.unwrap().0, valid_child.block_hash());
     }
 }
